0.0.68 1.修改科技个位数的BUG 2.修改白马和骑射不生效的问题 3.其他一些大家上报的BUG 4. 疾驰特技实装

+++ Project/Assets/Sango/Scripts/UI/Common/UISortButton.cs
@@ -26,7 +26,7 @@ using Sango.Core; namespace Sango.UI
             nameLab.text = name;
             return this;
         }
-        public UISortButton SetWidth(int width)
+        public UISortButton SetWidth(float width)
         {
             RectTransform rectTransform = GetComponent<RectTransform>();
             rectTransform.sizeDelta = new Vector2(width, rectTransform.sizeDelta.y);
